+++ test/TestInterval.hs
@@ -17,6 +17,9 @@ import Test.Tasty
 import Test.Tasty.QuickCheck
 import Test.Tasty.HUnit
 import Test.Tasty.TH
+#ifdef MIN_VERSION_quickcheck_classes_base
+import Test.QuickCheck.Classes.Base
+#endif
 
 import Data.Interval
   ( Interval, Extended (..), (<=..<=), (<=..<), (<..<=), (<..<)
@@ -927,6 +930,17 @@ case_Data = everywhere f i @?= (1 <=..<= 2 :: Interval Integer)
     f x
       | Just (y :: Integer) <- cast x = fromJust $ cast (y + 1)
       | otherwise = x
+
+{--------------------------------------------------------------------
+  Storable
+--------------------------------------------------------------------}
+
+#ifdef MIN_VERSION_quickcheck_classes_base
+test_Storable = map (uncurry testProperty) $ lawsProperties $
+  storableLaws (Proxy :: Proxy (Interval Double))
+#else
+test_Storable = []
+#endif
 
 {--------------------------------------------------------------------
   Generators