+++ CHANGELOG.markdown
@@ -3,6 +3,7 @@
 
 * introduce operations for Allen's interval algebra (#18, thanks to marcosh)
 * amend `recip` so that `recip (recip xs) == xs // {0}` (#21)
+* add `instance Storable` for `Interval` (#25)
 
 2.0.0
 -----